Refactor internals to make data flow and logic clearer.
There was a lot of mess in how container_lifecycle and serve
were interacting, often specifically caused by the containerconfig
type that I introduced a while ago. It had names for variables
that did not align correctly with the container.Config and
container.HostConfig types, meaning there was cognitive overhead
that didn't need to be there.

Everything is much simpler now.