Exposes a plugins command that copies files out of the container.
This is coming to grips with the first part of the required functionality.

Of all the approaches I thought of, I kept thinking that my approach to
generating plugins.txt should maintain the container as stateless.

Because of this, I didn't want to copy a helper binary into the container
at runtime.

I don't want to introduce a jinx-specific helper built into the container,
either, because then I need to update and manage the helper at the same time as
I update and manage everything else.

Instead, I'm choosing to copy information from the container into the userland
of the host, from where I can calculate it as I please.