Only expose 'new' subcommand when no jiinx project is detected. Fix subcommands/options.
Only exposing new seems non-ideal, but I introduced a hard dependency
in the subcommands on the config file when those commands are registered.

I should tweak that away, but this works for now and lets me make progress.

The subcommands fixing was silly; I've been adding the subcommands before
I set them up, so of course they did not work.